Update code coverage.
Current number has it at 75% but I think the postgrex types are
effecting the number. Removed the migration file and placed the
migration inside of test_helper.exs so that it did not have
an effect on coverage either. For all relevant code, there are
no lines missing from what can be seen from coverage report.

connects to #2